Move Checks class to proper package
