+++ src/main/java/org/toradocu/translator/Subject.java
@@ -7,7 +7,7 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.Objects;
 import java.util.stream.Collectors;
-import org.toradocu.Checks;
+import org.toradocu.util.Checks;
 
 /**
  * Represents a subject in a sentence. The subject contains also the information about the container
