+++ src/main/java/org/toradocu/util/RandoopSpecs.java
@@ -4,7 +4,6 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
 import java.util.stream.Collectors;
-import org.toradocu.Checks;
 import org.toradocu.conf.Configuration;
 import org.toradocu.extractor.ExecutableMember;
 import org.toradocu.extractor.ParamTag;
